+++ tests/test_api.py
@@ -10,12 +10,13 @@ def test_kraken_api(get_client, real):
 def test_kraken_get_account_balance(get_client, real):
     client = get_client(real)
     balances = client.get_account_balance()
+    print(balances)
     assert balances is not None
 
 
 def test_kraken_get_order_book(get_client, real):
     client = get_client(real)
-    order_book = client.get_order_book((Currency.BTC, Currency.USD))
+    order_book = client.get_order_book((Currency.XBT, Currency.USD))
     assert order_book is not None
     assert "asks" in order_book.keys()
     assert "bids" in order_book.keys()
@@ -30,12 +31,23 @@ def test_kraken_get_closed_orders(get_client, real):
 
 def test_kraken_get_ticker_info(get_client, real):
     client = get_client(real)
-    btc_ticker_info = client.get_ticker_info((Currency.BTC, Currency.USD))
+    btc_ticker_info = client.get_ticker_info((Currency.XBT, Currency.USD))
 
     assert isinstance(btc_ticker_info[0], TickerInfo)
-    assert btc_ticker_info[0].currency_pair_id == "XBTZUSD"
+    assert btc_ticker_info[0].currency_pair_id == "XBTUSD"
 
     usdt_ticker_info = client.get_ticker_info((Currency.USDT, Currency.USD))
 
     assert isinstance(usdt_ticker_info[0], TickerInfo)
     assert usdt_ticker_info[0].currency_pair_id == "USDTUSD"
+
+    eth_ticker_info = client.get_ticker_info((Currency.ETH, Currency.USDT))
+
+    assert isinstance(eth_ticker_info[0], TickerInfo)
+    assert eth_ticker_info[0].currency_pair_id == "ETHUSDT"
+
+    # reversed, to check if that works as well
+    eth_ticker_info2 = client.get_ticker_info((Currency.USDT, Currency.ETH))
+
+    assert isinstance(eth_ticker_info2[0], TickerInfo)
+    assert eth_ticker_info2[0].currency_pair_id == "ETHUSDT"